test: update videos controller tests
